fix declaration
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@31322 72102866-910b-0410-8b05-ffd578937521

+++ Headers/Foundation/NSPort.h
@@ -249,7 +249,7 @@ typedef SOCKET NSSocketNativeHandle;
  * This is a callback method used by the NSRunLoop class to determine which
  * descriptors to watch for the port.
  */
-- (void) getFds: (int*)fds count: (int*)count;
+- (void) getFds: (NSInteger*)fds count: (NSInteger*)count;
 
 /**
  *  Delegates processing of a message.
